refactor: Decompose text editor and text editor wrapper
